[REFACTOR] `fileSystem.js` code moved to `abstractFilesystemAccess.js` [REFACTOR] github handle updated
