+++ tests/e2e/spec/gutenberg_ro.e2e.spec.js
@@ -1,7 +1,7 @@
 /**
  * legacy-ray_charles.e2e.spec.js : End-to-end tests implemented with Selenium WebDriver and Mocha
  *
- * Copyright 2023 Jaifroid, RG7279805 and contributors
+ * Copyright 2023 Jaifroid, Rishabhg71 and contributors
  * Licence GPL v3:
  *
  * This file is part of Kiwix.
@@ -192,7 +192,6 @@ function runTests (driver, modes) {
                     const archiveFiles = await driver.findElement(By.id('archiveFiles'));
                     if (!isFileLoaded) await archiveFiles.sendKeys(gutenbergRoBaseFile);
                     filesLength = await driver.executeScript('return document.getElementById("archiveFiles").files.length');
-
                     // In new browsers Files are loaded using the FileSystem API, so we have to set the local archives using JavaScript
                     // which were selected using the file input
                     await driver.executeScript('window.setLocalArchiveFromFileSelect();');
